requirements.txt
Django==4.0.3
Pillow==9.0.1
djangorestframework==3.13.1
